requirements.txt
streamlit
spacy
python-docx
PyPDF2
scikit-learn
requests
fpdf